+++ contentstack/asset.py
@@ -22,7 +22,8 @@ class Asset:
         self.__uid = uid
         if self.__uid is None or self.__uid.strip() == 0:
             raise KeyError('Please provide valid uid')
-        self.base_url = '{}/assets/{}'.format(self.http_instance.endpoint, self.__uid)
+        self.base_url = '{}/assets/{}'.format(
+            self.http_instance.endpoint, self.__uid)
         if 'environment' in self.http_instance.headers:
             self.asset_params['environment'] = self.http_instance.headers['environment']