+++ contentstack/image_transform.py
@@ -13,6 +13,7 @@ import logging
 
 log = logging.getLogger(__name__)
 
+
 class ImageTransform:  # pylint: disable=too-few-public-methods
     """
     The Image Delivery API is used to retrieve, manipulate and/or convert image
@@ -50,8 +51,8 @@ class ImageTransform:  # pylint: disable=too-few-public-methods
             >>> result = image_url.fetch()
         ------------------------------
         """
-        # args = ['{0}={1}'.format(k, v) for k, v in self.image_params.items()]
-        args = ['{0}={1}'.format(k, v) for k, v in list(self.image_params.items())]
+        args = ['{0}={1}'.format(k, v)
+                for k, v in list(self.image_params.items())]
         if args:
             self.image_url += '?{0}'.format('&'.join(args))
         return self.image_url