+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 # CHANGELOG
 
+## _v1.6.0_
+
+============
+
+### **Date: 11-Aug-2021**
+
+Live Preview support added.
+
+- Stack.live_preview_query function added in Stack
+- live preview funtions added in config
+
 ## _v1.5.1_
 
 ============
 
-**Date: 1-Aug-2021**
+### **Date: 1-Aug-2021**
 
 Issue #17 resolved.
 Stack.sync_init uses wrong parameter names
@@ -13,7 +24,7 @@ Stack.sync_init uses wrong parameter names
 
 ============
 
-**Date: 22-Jul-2021**
+### **Date: 22-Jul-2021**
 
 contentstack-utils updated to v1.1.0
 
@@ -21,7 +32,7 @@ contentstack-utils updated to v1.1.0
 
 ============
 
-**Date: 05-Apr-2021**
+### **Date: 05-Apr-2021**
 
 Entry
     - include_embedded_objects support added