+++ changelog.rst
@@ -2,6 +2,15 @@
 **CHANGELOG**
 ================
 
+*v1.6.0*
+============
+
+**Date: 11-Aug-2021**
+
+Live Preview support added.
+- Stack.live_preview_query function added in Stack
+- live preview funtions added in config
+
 *v1.5.1*
 ============